move aiHisto constants to shared/utils instead of shared/types, to prevent cyclic imports
