+++ server/routes/aiProjectSelectedWSImages.ts
@@ -1,11 +1,7 @@
 import { type Annotation, type Mds3, type RouteApi, type WSImage } from '#types'
 import type { AiProjectSelectedWSImagesRequest, AiProjectSelectedWSImagesResponse, FlagPredictionInfo } from '#types'
-import {
-	aiProjectSelectedWSImagesResponsePayload,
-	createSelectionID,
-	SelectionPrefixes,
-	FlagStatus
-} from '#types/checkers'
+import { aiProjectSelectedWSImagesResponsePayload } from '#types/checkers'
+import { createSelectionID, SelectionPrefixes, FlagStatus } from '#shared'
 import { getDbConnection } from '#src/aiHistoDBConnection.ts'
 import type Database from 'better-sqlite3'
 /*
