+++ shared/types/checkers/aiProjectSelectedWSImages.ts
@@ -5,14 +5,7 @@ import type {
 	AiProjectSelectedWSImagesResponse
 } from '../src/routes/aiProjectSelectedWSImages.ts'
 
-export { FeaturePrefixes } from '../src/routes/aiProjectSelectedWSImages.ts'
-export { FlagStatus } from '../src/routes/aiProjectSelectedWSImages.ts'
-export { FlagStatusMessages } from '../src/routes/aiProjectSelectedWSImages.ts'
-export { SelectionPrefixes } from '../src/routes/aiProjectSelectedWSImages.ts'
 export { aiProjectSelectedWSImagesResponsePayload } from '../src/routes/aiProjectSelectedWSImages.ts'
-export { checkSelectionType } from '../src/routes/aiProjectSelectedWSImages.ts'
-export { createFeatureID } from '../src/routes/aiProjectSelectedWSImages.ts'
-export { createSelectionID } from '../src/routes/aiProjectSelectedWSImages.ts'
 
 export const validAiProjectSelectedWSImagesRequest = createValidate<AiProjectSelectedWSImagesRequest>()
 export const validAiProjectSelectedWSImagesResponse = createValidate<AiProjectSelectedWSImagesResponse>()
