+++ shared/types/package.json
@@ -11,7 +11,8 @@
     "./checkers": "./dist/index.js"
   },
   "imports": {
-    "#shared": "@sjcrh/proteinpaint-shared"
+    "#shared": "@sjcrh/proteinpaint-shared",
+    "#shared/constants/*": "@sjcrh/proteinpaint-shared/constants/*"
   },
   "scripts": {
     "generate": "node dev.mjs",
