+++ shared/types/src/routes/aiProjectSelectedWSImages.ts
@@ -1,5 +1,7 @@
 import type { RoutePayload } from './routeApi.ts'
 import type { WSImage } from './samplewsimages.ts'
+import type { FlagStatus } from '#shared/constants/AiHisto.d.ts'
+
 export type AiProjectSelectedWSImagesRequest = {
 	genome: string
 	dslabel: string
@@ -10,44 +12,6 @@ export type AiProjectSelectedWSImagesRequest = {
 export type AiProjectSelectedWSImagesResponse = {
 	// TODO create a type for WSImage with AI project specific fields
 	wsimages: WSImage[]
-}
-
-export enum FlagStatus {
-	Normal = 0,
-	Skipped = 1,
-	Flagged = 2,
-	Deleted = 3
-}
-
-export enum FeaturePrefixes {
-	Star = 'annotation-star-',
-	Square = 'annotation-square-',
-	Border = 'annotation-border-',
-	PredBorder = 'prediction-border-'
-}
-
-export enum SelectionPrefixes {
-	TileSelection = 'ts_',
-	Prediction = 'pred_',
-	Annotation = 'anno_'
-}
-
-export const FlagStatusMessages = {
-	[FlagStatus.Normal]: '',
-	[FlagStatus.Skipped]: '(Skipped)',
-	[FlagStatus.Flagged]: '(Flagged)'
-}
-
-export function createSelectionID(prefix: SelectionPrefixes, coordinates: [number, number]): string {
-	return prefix + JSON.stringify(coordinates)
-}
-
-export function checkSelectionType(tileSelection: TileSelection, suspectedPrefix: SelectionPrefixes): boolean {
-	return tileSelection.id.startsWith(suspectedPrefix)
-}
-
-export function createFeatureID(featurePrefix: FeaturePrefixes, coords: [number, number]) {
-	return featurePrefix + JSON.stringify(coords)
 }
 
 export interface FlagPredictionInfo {
