+++ shared/utils/.gitignore
@@ -1,1 +1,2 @@
 dist/*
+constants/*.d.ts
