+++ shared/utils/package.json
@@ -10,17 +10,19 @@
     "#types/checkers": "@sjcrh/proteinpaint-types/checkers"
   },
   "exports": {
-    ".": "./dist/index.js",
-    "./*.ts": "./dist/*.ts_SHOULD_BE_js",
-    "./*": "./dist/*"
+    ".": "./dist/src/index.js",
+    "./*.ts": "./*.ts_SHOULD_BE_js",
+    "./*": "./dist/src/*",
+    "./constants/*": "./dist/constants/*"
   },
   "scripts": {
-    "build": "esbuild src/*.* --platform=node --outdir=dist --format=esm --sourcemap",
-    "prepack": "npm run build && tsc --sourcemap",
+    "build": "esbuild src/*.* constants/*.ts --platform=node --outdir=dist --format=esm --sourcemap",
+    "prepack": "rm -rf dist && npm run build && tsc --sourcemap --preserveWatchOutput",
+    "dev": "rm -rf dist && concurrently 'npm run build -- --watch' 'npx tsc --watch --skipLibCheck --preserveWatchOutput'",
     "pretest": "mkdir -p test && node emitImports > test/internals-test.ts",
-    "spec:coverage": "node test/relevant.js",
     "test": "tsx test/internals-test.ts",
-    "test-x": "ls src/test/*.spec* | xargs -I % bash -c '{ tsx %; sleep 0.001; }'"
+    "test-x": "ls src/test/*.spec* | xargs -I % bash -c '{ tsx %; sleep 0.001; }'",
+    "spec:coverage": "node test/relevant.js"
   },
   "author": "",
   "license": "ISC",
